Update 1_readwritebyte.s
Changed the way the SPI is handeled. instead of bitbanging, the SPI out is now handled by the internal shift register of the 6522. For reading the incoming bytes, a 74hc595 is used witch is connected to PORTA of the VIA. The SD_CS signal is controlled by
pin 7 of PORTB.